patch to hide some vm's from -a
There are some vm's that don't need to show up in -a when not running,
but the current -r option blasts all vm's that aren't running, which
is more than I want.

This patch allows adding an "ign":true to the machine-enviornments
object for each path that shouldn't be displayed in -a when not running.